refactor(pdf): remove deprecated trial dog PDF phase 3 and update rendering logic
- Deleted the unused `drawTrialDogPdfPhase3` function.
- Updated the PDF rendering logic to utilize new internal helper functions for drawing trial dog details and background.
- Added new test files for the internal helper functions `drawTrialDogPdfKoiranTiedot` and `drawTrialDogPdfKokeenTiedot`.
- Removed the old test file for phase 2 as part of the refactor.

Refs: BEJ-89
