+++ apps/web/lib/server/trials/internal/koiran-tausta.ts
@@ -1,0 +1,6 @@
+import type { PDFPage } from "pdf-lib";
+
+export function drawTrialDogPdfKoiranTausta(input: { page: PDFPage }): void {
+  void input;
+  // Koiran tausta will be added here as the next internal helper block.
+}
