--- apps/web/lib/server/trials/trial-dog-pdf-phase-3.ts
@@ -1,6 +1,0 @@
-import type { PDFPage } from "pdf-lib";
-
-export function drawTrialDogPdfPhase3(input: { page: PDFPage }): void {
-  void input;
-  // Phase 3 will cover the parent and owner block once those fields are wired in.
-}
